Moved TextContainer and TranslationContainer from event to lang
