+++ src/main/java/cn/nukkit/command/CommandSender.java
@@ -1,7 +1,7 @@
 package cn.nukkit.command;
 
 import cn.nukkit.Server;
-import cn.nukkit.event.TextContainer;
+import cn.nukkit.lang.TextContainer;
 import cn.nukkit.permission.Permissible;
 
 /**
