+++ src/main/java/cn/nukkit/command/ConsoleCommandSender.java
@@ -1,7 +1,7 @@
 package cn.nukkit.command;
 
 import cn.nukkit.Server;
-import cn.nukkit.event.TextContainer;
+import cn.nukkit.lang.TextContainer;
 import cn.nukkit.permission.PermissibleBase;
 import cn.nukkit.permission.Permission;
 import cn.nukkit.permission.PermissionAttachment;
