+++ src/main/java/cn/nukkit/command/PluginCommand.java
@@ -1,6 +1,6 @@
 package cn.nukkit.command;
 
-import cn.nukkit.event.TranslationContainer;
+import cn.nukkit.lang.TranslationContainer;
 import cn.nukkit.plugin.Plugin;
 
 /**
