+++ src/main/java/cn/nukkit/command/defaults/BanCommand.java
@@ -3,7 +3,7 @@ package cn.nukkit.command.defaults;
 import cn.nukkit.Player;
 import cn.nukkit.command.Command;
 import cn.nukkit.command.CommandSender;
-import cn.nukkit.event.TranslationContainer;
+import cn.nukkit.lang.TranslationContainer;
 
 import java.util.Objects;
