+++ src/main/java/cn/nukkit/command/defaults/DefaultGamemodeCommand.java
@@ -2,7 +2,7 @@ package cn.nukkit.command.defaults;
 
 import cn.nukkit.Server;
 import cn.nukkit.command.CommandSender;
-import cn.nukkit.event.TranslationContainer;
+import cn.nukkit.lang.TranslationContainer;
 
 /**
  * Created on 2015/11/12 by xtypr.
