+++ src/main/java/cn/nukkit/command/defaults/GiveCommand.java
@@ -3,7 +3,7 @@ package cn.nukkit.command.defaults;
 import cn.nukkit.Player;
 import cn.nukkit.command.Command;
 import cn.nukkit.command.CommandSender;
-import cn.nukkit.event.TranslationContainer;
+import cn.nukkit.lang.TranslationContainer;
 import cn.nukkit.item.Item;
 import cn.nukkit.utils.TextFormat;
