+++ src/main/java/cn/nukkit/command/defaults/SeedCommand.java
@@ -2,7 +2,7 @@ package cn.nukkit.command.defaults;
 
 import cn.nukkit.Player;
 import cn.nukkit.command.CommandSender;
-import cn.nukkit.event.TranslationContainer;
+import cn.nukkit.lang.TranslationContainer;
 
 /**
  * author: MagicDroidX
