+++ src/main/java/cn/nukkit/command/defaults/TimingsCommand.java
@@ -1,7 +1,7 @@
 package cn.nukkit.command.defaults;
 
 import cn.nukkit.command.CommandSender;
-import cn.nukkit.event.TranslationContainer;
+import cn.nukkit.lang.TranslationContainer;
 import cn.nukkit.timings.Timings;
 import cn.nukkit.timings.TimingsExport;
