+++ src/main/java/cn/nukkit/event/player/PlayerDeathEvent.java
@@ -3,7 +3,7 @@ package cn.nukkit.event.player;
 import cn.nukkit.Player;
 import cn.nukkit.event.Cancellable;
 import cn.nukkit.event.HandlerList;
-import cn.nukkit.event.TextContainer;
+import cn.nukkit.lang.TextContainer;
 import cn.nukkit.event.entity.EntityDeathEvent;
 import cn.nukkit.item.Item;
