+++ src/main/java/cn/nukkit/event/player/PlayerJoinEvent.java
@@ -2,7 +2,7 @@ package cn.nukkit.event.player;
 
 import cn.nukkit.Player;
 import cn.nukkit.event.HandlerList;
-import cn.nukkit.event.TextContainer;
+import cn.nukkit.lang.TextContainer;
 
 public class PlayerJoinEvent extends PlayerEvent {
     private static final HandlerList handlers = new HandlerList();
