+++ src/main/java/cn/nukkit/lang/BaseLang.java
@@ -1,8 +1,6 @@
 package cn.nukkit.lang;
 
 import cn.nukkit.Server;
-import cn.nukkit.event.TextContainer;
-import cn.nukkit.event.TranslationContainer;
 import cn.nukkit.utils.Utils;
 
 import java.io.IOException;
