+++ src/main/java/cn/nukkit/lang/TextContainer.java
@@ -1,4 +1,4 @@
-package cn.nukkit.event;
+package cn.nukkit.lang;
 
 import cn.nukkit.Server;
