+++ src/main/java/cn/nukkit/lang/TranslationContainer.java
@@ -1,4 +1,4 @@
-package cn.nukkit.event;
+package cn.nukkit.lang;
 
 /**
  * author: MagicDroidX
